+++ test/tools/format_tool_test.dart
@@ -58,6 +58,7 @@ void main() {
             formatterInputs.includedFiles,
             unorderedEquals({
               'file.dart',
+              'links/not_link.dart',
               'lib/sub/file.dart',
               'linked.dart',
               'other/file.dart',
@@ -69,6 +70,35 @@ void main() {
             unorderedEquals({'.dart_tool_test'}));
         expect(formatterInputs.skippedLinks,
             unorderedEquals({'links/lib-link', 'links/link.dart'}));
+      });
+
+      test('custom excludes with collapseDirectories', () {
+        FormatterInputs formatterInputs = FormatTool.getInputs(
+          exclude: [Glob('*_exclude.dart')],
+          root: root,
+          collapseDirectories: true,
+        );
+
+        expect(
+          formatterInputs.includedFiles,
+          unorderedEquals({
+            'file.dart',
+            'lib',
+            'linked.dart',
+            'other',
+            'links',
+          }),
+        );
+
+        expect(
+          formatterInputs.excludedFiles,
+          unorderedEquals({'should_exclude.dart'}),
+        );
+        expect(
+          formatterInputs.hiddenDirectories,
+          unorderedEquals({'.dart_tool_test'}),
+        );
+        expect(formatterInputs.skippedLinks, isEmpty);
       });
 
       test('empty inputs due to excludes config', () async {
@@ -85,6 +115,7 @@ void main() {
             formatterInputs.includedFiles,
             unorderedEquals({
               'file.dart',
+              'links/not_link.dart',
               'lib/sub/file.dart',
               'linked.dart',
               'other/file.dart',
@@ -100,6 +131,7 @@ void main() {
             formatterInputs.includedFiles,
             unorderedEquals({
               'lib-link/sub/file.dart',
+              'not_link.dart',
               'link.dart',
             }));
         expect(formatterInputs.skippedLinks, isEmpty);
@@ -199,16 +231,14 @@ void main() {
       expect(execution.exitCode, ExitCode.config.code);
     });
 
-    test('logs the excluded paths, skipped links and hidden directories',
-        () async {
-      Logger.root.level = Level.ALL;
+    test('logs the excluded paths and hidden directories', () async {
+      var currentLevel = Logger.root.level;
+      Logger.root.level = Level.FINE;
       expect(
           Logger.root.onRecord,
           emitsInOrder([
             fineLogOf(allOf(contains('Excluding these paths'),
                 contains('should_exclude.dart'))),
-            fineLogOf(allOf(contains('Excluding these links'),
-                contains('lib-link'), contains('link.dart'))),
             fineLogOf(allOf(contains('Excluding these hidden directories'),
                 contains('.dart_tool_test'))),
           ]));
@@ -216,6 +246,25 @@ void main() {
       buildExecution(DevToolExecutionContext(),
           exclude: [Glob('*_exclude.dart')],
           path: 'test/tools/fixtures/format/globs');
+
+      Logger.root.level = currentLevel;
+    });
+
+    test('logs the skipped links', () async {
+      var currentLevel = Logger.root.level;
+      Logger.root.level = Level.FINE;
+      expect(
+          Logger.root.onRecord,
+          emitsInOrder([
+            fineLogOf(allOf(contains('Excluding these links'),
+                contains('lib-link'), contains('link.dart'))),
+          ]));
+
+      buildExecution(DevToolExecutionContext(),
+          exclude: [Glob('*_exclude.dart')],
+          path: 'test/tools/fixtures/format/globs/links');
+
+      Logger.root.level = currentLevel;
     });
 
     group('returns a FormatExecution', () {